+++ bzxt117/apps/sample/views.py
@@ -265,7 +265,10 @@ class exploSampleGCMSTestFileViewset(viewsets.ModelViewSet):
             #         d = os.path.join(dirPath, file)
             #         fileDel = exploSampleGCMSTestFile.objects.filter(txtURL = d)
             #         fileDel.delete()
-            shutil.rmtree(filePath)
+            if os.path.exists(filePath) == True:
+                shutil.rmtree(filePath)
+            else:
+                raise APIException("想要删除的GC_MS文件夹路径不存在")
         instance.delete()
 
 class devSampleViewset(viewsets.ModelViewSet):
